refactor: acquire

+++ PlatformIO/phmeter/include/acquire.h
@@ -12,7 +12,7 @@ void runExperiment(uint8_t);
 void runExperiment();
 void runSequence(uint8_t);
 void calculateResult(uint8_t);
-long acquireOne(uint8_t);
+uint16_t acquireOne(uint8_t);
 void acquire(bool);
 void printColor(Print*, uint8_t);
 void printColorOne(Print*, uint8_t);
